refactor(core): extract maintenance/reliability.ts — getReliabilityMetrics + rawEventsGate
